producer/requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
iniconfig==2.0.0
kafka-python==2.0.6
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytz==2025.1
requests==2.32.3
urllib3==2.3.0
xmltodict==0.14.2
